media_processing_agent/functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.1.0
google-cloud-firestore>=2.21.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
python-dotenv>=1.1.1